refactor tensor module
